renamed data_source_registry to be more clear
